docs: update layout pipe demmo
