+++ packages/g6/package.json
@@ -66,7 +66,7 @@
     ]
   },
   "dependencies": {
-    "@antv/g6-pc": "*"
+    "@antv/g6-pc": "0.3.2"
   },
   "devDependencies": {
     "@babel/core": "^7.7.7",
