+++ packages/pc/package.json
@@ -74,9 +74,9 @@
     "@antv/g-canvas": "^0.5.2",
     "@antv/g-math": "^0.1.1",
     "@antv/g-svg": "^0.5.1",
-    "@antv/g6-core": "*",
-    "@antv/g6-plugin": "*",
-    "@antv/g6-element": "*",
+    "@antv/g6-core": "0.3.2",
+    "@antv/g6-plugin": "0.3.2",
+    "@antv/g6-element": "0.3.2",
     "@antv/algorithm": "^0.1.8",
     "@antv/hierarchy": "^0.6.7",
     "@antv/layout": "^0.1.14",
