+++ packages/plugin/package.json
@@ -22,7 +22,7 @@
     "@antv/g-base": "^0.5.1",
     "@antv/g-canvas": "^0.5.2",
     "@antv/g-svg": "^0.5.2",
-    "@antv/g6-core": "*",
+    "@antv/g6-core": "0.3.2",
     "@antv/matrix-util": "^3.0.4",
     "@antv/scale": "^0.3.4",
     "@antv/util": "^2.0.9",
