+++ packages/site/examples/tool/legend/demo/meta.json
@@ -19,6 +19,14 @@
         "en": "Legend with Hovering Filtering"
       },
       "screenshot": "https://gw.alipayobjects.com/mdn/rms_f8c6a0/afts/img/A*ftAtR6k9RWIAAAAAAAAAAAAAARQnAQ"
+    },
+    {
+      "filename": "legendClick.js",
+      "title": {
+        "zh": "图例-点击筛选2",
+        "en": "Legend with Clicking Filtering 2"
+      },
+      "screenshot": "https://gw.alipayobjects.com/mdn/rms_f8c6a0/afts/img/A*nCTfRq6522kAAAAAAAAAAAAAARQnAQ"
     }
   ]
 }
